+++ tools/extendedInteractiveLesson/tool.js
@@ -57,7 +57,13 @@
         'extendedInteractiveOptionsTitle': 'LiChess Tools - interactive lesson preferences',
         'giveUpButtonText': 'Give up',
         'giveUpButtonTitle': 'Abandons the interactive run',
-        'giveUpConfirmation': 'Are you sure you want to abandon the interactive run?'
+        'giveUpConfirmation': 'Are you sure you want to abandon the interactive run?',
+        'daysText:one': 'a day',
+        'hoursText:one': 'an hr',
+        'minutesText:one': 'a min',
+        'daysText': '%s days',
+        'hoursText': '%s hrs',
+        'minutesText': '%s mins'
       },
       'ro-RO': {
         'options.study': 'Studiu',
@@ -93,7 +99,13 @@
         'extendedInteractiveOptionsTitle': 'LiChess Tools - preferin\u0163e lec\u0163ie interactiv\u0103',
         'giveUpButtonText': 'Renun\u0163',
         'giveUpButtonTitle': 'Abandoneaz\u0103 lec\u0163ia interactiv\u0103',
-        'giveUpConfirmation': 'E\u015Fti sigur ca vrei sa abandonezi lec\u0163ia interactiv\u0103?'
+        'giveUpConfirmation': 'E\u015Fti sigur ca vrei sa abandonezi lec\u0163ia interactiv\u0103?',
+        'daysText:one': 'o zi',
+        'hoursText:one': 'o or\u0103',
+        'minutesText:one': 'un minut',
+        'daysText': '%s zile',
+        'hoursText': '%s ore',
+        'minutesText': '%s minute'
       }
     }
 
@@ -122,7 +134,7 @@
             gp.currentPath = this.getCurrentPath();
             if (!gp.currentPath) {
               const nextMoves = lt.getNextMoves(node, gp.threeFoldRepetition)
-                .filter(c => this.isPermanentNode(c));
+                .filter(c => this.isPermanentNode(c) && !this.areBadGlyphNodes([c]));
               if (nextMoves.length) {
                 if (lt.global.confirm(trans.noarg('resetQuestionNoVariations'))) { //TODO can we make this await and use uiApi.dialog?
                   this.resetDone();
@@ -466,23 +478,8 @@
         return elem;
       };
 
-      const colorCodeFactor = (factor) => {
-        const cls = factor >= 1 ? 'green'
-                     : factor >= 0.85 ? 'yellow'
-                     : factor >= 0.7 ? 'orange'
-                     : 'red';
-        const elem = $('<span>')
-                 .addClass('lichessTools-rate-'+cls)
-                 .text((factor*100).toFixed(0)+'%');
-        return elem;
-      };
-
       const previousInterval = item.interval;
-      let infoElem = $('.lichessTools-extendedInteractiveLesson-info');
-      if (!infoElem.length) {
-        infoElem = $('<div class="lichessTools-extendedInteractiveLesson-info">')
-          .appendTo('.gamebook .comment');
-      }
+      const infoElem = $('<div class="lichessTools-extendedInteractiveLesson-info">');
       $('<div>')
        .appendSpan(`${goodMoves} ${lt.icon.Checked} | ${badMoves} ${lt.icon.RedX} (`)
        .append(colorCodeSuccessRate(successRate))
@@ -495,11 +492,19 @@
           : 0.5 + 0.5 * (successRate / 0.7);
         item.interval = Math.max(1/144, item.interval * factor);
         $('<div>')
-          .appendSpan(trans.pluralSame('daysText',`${previousInterval.toFixed(1)} ${lt.icon.RightwardsArrow} ${item.interval.toFixed(1)}`) +' (')
-          .append(colorCodeFactor(factor))
-          .appendSpan(')')
+          .text(trans.pluralSame('daysText',`${previousInterval.toFixed(1)} ${lt.icon.RightwardsArrow} ${item.interval.toFixed(1)}`))
           .appendTo(infoElem);
       }
+      const attach = ()=>{
+        const container = $('.gamebook .comment');
+        if (!container.length) {
+          lt.global.setTimeout(attach,100);
+          return;
+        }
+        $('.lichessTools-extendedInteractiveLesson-info',container).remove();
+        infoElem.appendTo(container);
+      };
+      attach();
 
       paths[path] = item;
 
@@ -1050,6 +1055,26 @@
       }
     };
 
+    getTimeText = (value) => {
+      const lt = this.lichessTools;
+      const trans = lt.translator;
+      let result;
+      const days = Math.round(value / 86400000);
+      if (Math.trunc(value / 86400000)) {
+        result = trans.plural('daysText', days, days);
+      } else {
+        const hours = Math.round(value / 3600000);
+        if (Math.trunc(value / 3600000)) {
+          result = trans.plural('hoursText', hours, hours);
+        } else {
+          const minutes = Math.round(value / 60000);
+          result = trans.plural('minutesText', minutes, minutes)
+        }
+      }
+      return result;
+    };
+
+
     refreshChapterProgress = () => {
       const lt = this.lichessTools;
       const lichess = lt.lichess;
@@ -1059,6 +1084,8 @@
       if (!study) return;
       const trans = lt.translator;
       this.loadChapterPaths(null);
+      //$('.study__chapters button[title]').removeAttr('title');
+      $('.study__chapters button[data-tooltip]').removeAttr('data-tooltip');
       if (!this._paths) return;
       const list = study.chapters.list.all();
       $('div.study__chapters').addClass('lichesstools-extendedInteractiveLessonFlow');
@@ -1071,20 +1098,31 @@
         let total = 0;
         let doneCount = 0;
         if (paths) {
+          let tooltip = '';
           for (const k in paths) {
             if (k == 'currentPath') continue;
             const item = paths[k];
             const done = this.options.flow.spacedRepetition
               ? item && Date.now() < item.time + item.interval * 86400000
               : item?.success
+            if (total<10) {
+              tooltip += '\u000a'+(total+1)+': '+(done?lt.icon.Checked:'')+(this.options.flow.spacedRepetition?' '+this.getTimeText(item.interval * 86400000):'');
+            } else if (total==10) {
+              tooltip += '\u000a  '+lt.icon.Ellipsis;
+            }
             total++;
             if (done) doneCount++;
           }
           if (total) {
+            tooltip = trans.pluralSame('progressTitle', doneCount + '/' + total)+tooltip;
             perc = (100 * doneCount / total) + '%';
-            container.attr('title', trans.pluralSame('progressTitle', doneCount + '/' + total));
+            container
+              //.attr('title', tooltip)
+              .attr('data-tooltip', tooltip);
           } else {
-            container.removeAttr('title');
+            container
+              //.removeAttr('title')
+              .removeAttr('data-tooltip');
           }
         }